+++ eslint.config.ts
@@ -22,6 +22,7 @@ export default antfu(
     rules: {
       'vue/block-order': ['error', { order: [['script', 'template'], 'style'] }],
       'vue/singleline-html-element-content-newline': ['off'],
+      'vue/component-api-style': ['warn'],
     },
   },
   {