+++ postcss.config.ts
@@ -1,16 +1,14 @@
 import postcssImport from 'postcss-import'
 import postcssPresetEnv from 'postcss-preset-env'
 import tailwindcss from 'tailwindcss'
+import type { Config } from 'postcss-load-config'
 
-// https://github.com/postcss/postcss#usage
-
-/** @type {import('postcss-load-config').Config} */
-export default {
+const config: Config = {
   plugins: [
     postcssImport(),
     tailwindcss(),
-    postcssPresetEnv({
-      stage: 1,
-    }),
+    postcssPresetEnv(),
   ],
 }
+
+export default config